api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "bcrypt>=4.3.0",
    "edgedb>=2.2.0",
    "fastapi[all]>=0.115.12",
    "gel>=3.1.0",
    "jinja2>=3.1.6",
    "pycryptodome>=3.23.0",
    "pydantic",
    "python-jose>=3.5.0",
]
